GatewayConnection: make m_receiveBuffer a local variable
There is no need to have the m_receiveBuffer in global scope
of the instance. We just need a private buffer while receiving
a message. Otherwise, there is a danger in manipulating with
a shared buffer with improper locking. We also reduce risks
with buffer reusing.

The #include <Poco/Buffer.h> was preserved as it will be used
soon in the future.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>