GatewayCommunicator: rate limit gateway registration
The method addGateway() is called when a new WebSocket connection from
a gateway comes with a registration message. This is out of control of
the rate limiter and thus an attacker can create a message loop of
creating new sockets sending new registration messages over and over
again.

This changes mostly covers this threat. Now, every addGateway() call
is also protected by a rateLimiter if any such appropriate one exists.
This means that an entirely new connection would not be limited but
any kind of reconnection (i.e. both reconnect and disconnect-connect)
is also rate limited.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>