GatewayCommunicator: handle multiple messages on readable
This effectively speeds up processing of incoming messages.
If there is too high trafiic from a single gateway, it is very
slow to always wait until the reactor schedules processing
on the per-message basis. We now do it in bulks.

The bulk processing is configurable. Each bulk has a maximum
number of messages it can process at once and the whole bulk
must be finished in a given time duration. If any of those
settings is reached, the bulk is finished until reactor calls
us again. Both maxMessageSize and maxBulkDuration are configurable
and can be disabled by setting to 0.

If the method receiveMessage() timeouts then we quit the bulk
processing as no more messages have come within quite a long time
period. Anyway, the receiveTimeout is changed to be shorter here
(1 s was too long).

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>