GatewayCommunicator: change logic of GatewayCommunicator and reactor
The GatewayConnection does not manage the SocketReactor and NObserver
anymore as this lead to use-after-free because of race conditions. The
issue was hidden inside the enqueueReadable(AutoPtr<ReadableNotification>)
method connected with the fact that NObserver does not held reference to
the GatewayConnection itself.

In case when the GatewayConnection was closed (being removed from the
m_connectionMap) and thus destroyed because nobody else owns a pointer
to it, if at the same time the appropriate socket comes from the reactor
to be served, the method enqueueReadable() would access deallocated
memory. This happens because enqueueReadable() touchs "self" (its
GatewayConnection via this) that is already deleted and not protected
by any smart pointer.

The only reasonable solution is to move the SocketReactor and NObserver
logic out of the GatewayConnection and manage the relation between
GatewayConnection and WebSocket explicitly inside GatewayCommunicator.

Unsuprisingly, this changes simplifies the code flow a bit, we have
to just take care extra about a new container m_socketsMap.

Fixes: 220e617aeeec ("GatewayConnection: new class")
Fixes: 36b6039833b1 ("GatewayCommunicator: new class")
Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>