t: xmlui: fix false positive when inviting as John
The session2 was created as John, however, we are incorrectly using
session of Joe. The false positive was probably occouring because
of timing issue. When admin invites a new admin but very close to
this operation, there was addition of another role, the time can
be too close and this might race. The reason is that either the
new admin role is created or not.

In this case, John (guest) is inviting a new admin which should
not work at all. But we should prevent the race condition anyway.

The error code 1001 means "forbidden" which is better and seems
to be correct. This test was probably written by somebody drunken...

Fixes: 4f81f860ed84 ("xmlui: t: test invite and remove")
Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>