GatewayRPCResult: introduce value SUCCESS_PARTIAL
The GatewayRPCResult can hold the terminal status SUCCESS_PARTIAL as
delivered from a GWResponse. Processing of such status is routed into
GatewayRPCHandler::onSuccess() and thus the client code would not
notice this change.

Operations that can success partially and need some extra care in that
case would have to be edited. There is currently no such operation.
User just would not be informed about the partiality.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>