DeviceServiceImpl: prefer DeviceDescription::name() over productName()
When a device is first registered, name it according to the DeviceDescription::name().
To stay backwards-compatible, if the name is missing, fallback to productName().
On update, use always only the DeviceDescription::name().

The GWS test test6_new_device_and_update must be updated accordingly. On update
of an existing device, the "product_name" is ignored, only "name" is considered.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>