DeviceServiceImpl: synchronize device properties
The DeviceServiceImpl now updates also device properties based on
details contained inside of the DeviceDescription. Not all contents
are synchronized, just the already supporty DevicePropertyKey values.

This commit however introduces a kind of an architectural mistake.
The DeviceServiceImpl should not have access to crypto config. Or,
only the GWSDeviceService methods should have that access. That is
because the GWSDeviceService includes also a part of the higher
level of the application (translation to domain objects).

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>