DeviceServiceImpl: check whether property exists first
If the property already exists, return false from the insert method.
This should be done in all the project later because the insert
conflicts are currently detected only via constraints violation
of the primary key which comes as an exception. However, the thrown
exception is of an unknown type (each database backend throws a
different exception). The generic DataException is not specific enough.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>